+++ sora-android-sdk/src/main/kotlin/jp/shiguredo/sora/sdk/channel/rtc/AndroidSystemCaSslCertificateVerifier.kt
@@ -1,6 +1,6 @@
 package jp.shiguredo.sora.sdk.channel.rtc
 
-import jp.shiguredo.sora.sdk.channel.tls.CustomCaTls
+import jp.shiguredo.sora.sdk.channel.tls.TlsConfigFactory
 import jp.shiguredo.sora.sdk.util.SoraLogger
 import org.webrtc.SSLCertificateVerifier
 import java.io.ByteArrayInputStream
@@ -25,7 +25,7 @@ internal class AndroidSystemCaSslCertificateVerifier(
         private const val OID_ANY_EXTENDED_KEY_USAGE = "2.5.29.37.0"
     }
 
-    private val trustManager: X509TrustManager = CustomCaTls.createTrustManager(caCertificate)
+    private val trustManager: X509TrustManager = TlsConfigFactory.createTrustManager(caCertificate)
 
     // verifyChain を実装している場合は verifyChain が呼び出されるため
     // verify は基本的に利用しない想定になっています。
